// src/pkts/localization/type.ts
export type InterfaceStrings = {
  title: {
    category: string;
    stories: string;
    availableDetails: string;
  };
  section: {
    historicalBackground: string;
    humanitarianImpact: string;
    didYouKnow: string;
    quote: string;
    highlights: string;
    perspectives: string;
    quickQuestions: string;
    technicalDetails: string;
    industryImpact: string;
    timeline: string;
    travelAdvisory: string;
    internationalReactions: string;
    actionItems: string;
    scientificSignificance: string;
    performanceStatistics: string;
    sources: string;
  };
  category: {
    ai: string;
    africa: string;
    apple: string;
    australia: string;
    austria: string;
    bay: string;
    belgium: string;
    bitcoin: string;
    brazil: string;
    business: string;
    canada: string;
    catholic: string;
    china: string;
    coffee: string;
    colombia: string;
    costa_rica: string;
    croatia: string;
    cryptocurrency: string;
    cybersecurity: string;
    czech_republic: string;
    denmark: string;
    economy: string;
    estonia: string;
    europe: string;
    finland: string;
    formula_1: string;
    france: string;
    gaming: string;
    germany_hesse: string;
    germany: string;
    greece: string;
    hungary: string;
    india: string;
    iran: string;
    ireland: string;
    israel: string;
    italy: string;
    japan: string;
    linux_oss: string;
    lithuania: string;
    mexico: string;
    middle_east: string;
    nfl: string;
    nhl: string;
    new_zealand: string;
    nintendo: string;
    norway: string;
    pakistan: string;
    palestine: string;
    philippines: string;
    podcasting: string;
    poland: string;
    portugal: string;
    romania: string;
    russia: string;
    science: string;
    serbia: string;
    singapore: string;
    slovakia: string;
    slovenia: string;
    south_korea: string;
    spain: string;
    sports: string;
    sweden: string;
    switzerland_de: string;
    switzerland_fr: string;
    taiwan: string;
    technology: string;
    thailand: string;
    the_netherlands: string;
    turkey: string;
    uk: string;
    ukraine: string;
    usa_austin: string;
    usa_chicago: string;
    usa_colorado: string;
    usa_michigan: string;
    usa_minnesota: string;
    usa_nyc: string;
    usa_ohio: string;
    usa_utah: string;
    usa_virginia: string;
    usa: string;
    iracing: string;
    world: string;
  };
};